Splitter htmlEditor fra redux-form-komponent
MELOSYS-4860
