+++ src/felleskomponenter/skjema/htmleditor/htmleditor.less
@@ -22,7 +22,13 @@
     margin: 0.5rem 0;
   }
 }
+
 .editor {
   padding-left: 0.5rem;
   padding-right: 0.5rem;
+}
+
+.editor_label {
+  margin-top: 1rem;
+  padding-bottom: 0.5rem;
 }
